requirements.txt
google-adk
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
beautifulsoup4
pypdf
pytest
reportlab
ddgs
